cnn_mnist_classifier/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli==1.27.38
flake8
python-dotenv>=0.5.1
pyaml>=5.3.1
